+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v.Next
 
 - `fetch` and `uncaught exception` instrumentation are included and enabled by default.
+- Resources Attributes now match React native environment
+- OS name and version now included in resource attributes
 
 ## v0.3.0